ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.35.0
groq==1.1.1
python-dotenv==1.2.1
pydantic==2.11.7
anyio==3.7.1
httpx==0.28.1